requirements.txt
python-dotenv
aiosmtplib
pytest
openpyxl
Jinja2